requirements.txt
selenium
fastapi
requests
uvicorn
pydantic_settings
webdriver_manager
psutil
argparse
typing 

pathlib 
langdetect
requests
googletrans
transformers
torch
opik
sentence_transformers
uptime
langchain_community
nltk
evaluate
google-api-python-client
accelerate
language_tool_python
lexical_diversity
sentencepiece
sarvamai
openpyxl

psycopg2-binary
mariadb
python-jose[cryptography]
passlib[bcrypt]
python-multipart
passlib>=1.7.4
bcrypt>=4.0.1
pydantic[email]
randomname
mariadb
gliner
python-iso639
ollama
deepeval
google-genai
rich
